Cambiada las rutas y test de user y tas terminado.

+++ src/main/java/com/todolist/controllers/UserController.java
@@ -24,7 +24,7 @@ import javax.validation.constraints.Pattern;
 import java.util.*;
 
 @RestController
-@RequestMapping("/api/v1/users")
+@RequestMapping("/api/v1")
 @Validated
 @AllArgsConstructor
 public class UserController {
@@ -36,8 +36,7 @@ public class UserController {
 
     private UserService userService;
 
-    @SuppressWarnings("unchecked")
-    @GetMapping
+    @GetMapping("/users")
     public List<Map<String, Object>> getAllUsers(@RequestParam(defaultValue = "0") @Min(value = 0, message = "The offset must be positive.") Integer offset,
                                                  @RequestParam(defaultValue = Integer.MAX_VALUE + "") @Min(value = 0, message = "The limit must be positive.") Integer limit,
                                                  @RequestParam(defaultValue = "idUser") String order,
@@ -79,8 +78,7 @@ public class UserController {
         return result.stream().map(user -> user.getFields(fieldsUser, fieldsTask)).toList();
     }
 
-    @SuppressWarnings("unchecked")
-    @GetMapping("/{idUser}")
+    @GetMapping("/user/{idUser}")
     public Map<String, Object> getUser(@PathVariable("idUser") @Min(value = 0, message = "The idUser must be positive.") Long idUser,
                                        @RequestParam(defaultValue = "idTask,title,description,status,finishedDate,startDate,annotation,priority,difficulty,duration") String fieldsTask,
                                        @RequestParam(defaultValue = "idUser,name,surname,email,avatar,bio,location,taskCompleted,tasks") String fieldsUser) {
@@ -94,7 +92,7 @@ public class UserController {
         return new ShowUser(user, userService.getShowTaskFromUser(user)).getFields(fieldsUser, fieldsTask);
     }
 
-    @PostMapping
+    @PostMapping("/user")
     public Map<String, Object> addUser(@RequestBody @Valid User user) {
         if (user == null)
             throw new BadRequestException("The user is null.");
@@ -114,7 +112,7 @@ public class UserController {
         return new ShowUser(user, userService.getShowTaskFromUser(user)).getFields(ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
 
-    @PutMapping
+    @PutMapping("/user")
     public Map<String, Object> updateUser(@RequestBody @Valid User user) {
         User oldUser = userService.findUserById(user.getIdUser());
         if (oldUser == null)
@@ -144,7 +142,7 @@ public class UserController {
         return new ShowUser(oldUser, userService.getShowTaskFromUser(oldUser)).getFields(ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
 
-    @DeleteMapping("/{idUser}")
+    @DeleteMapping("/user/{idUser}")
     public Map<String, Object> deleteUser(@PathVariable("idUser") @Min(value = 0, message = "The idGroup must be positive.") Long idUser) {
         User user = userService.findUserById(idUser);
         if (user == null)
@@ -153,9 +151,7 @@ public class UserController {
         return new ShowUser(user, userService.getShowTaskFromUser(user)).getFields(ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
 
-
-    @SuppressWarnings("unchecked")
-    @PostMapping("/{idUser}/tasks/{idTask}")
+    @PutMapping("/user/{idUser}/task/{idTask}")
     public Map<String, Object> addTaskToUser(@PathVariable("idUser") Long idUser, @PathVariable("idTask") Long idTask) {
         User user = userService.findUserById(idUser);
         if (user == null)
@@ -163,12 +159,11 @@ public class UserController {
         Task task = taskService.findTaskById(idTask);
         if (task == null)
             throw new NotFoundException("The task with idTask " + idTask + " does not exist.");
-        userService.addTaskToUser(user, task);
+        if (!userService.getTasksFromUser(user).contains(task)) userService.addTaskToUser(user, task);
         return new ShowUser(user, userService.getShowTaskFromUser(user)).getFields(ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
 
-    @SuppressWarnings("unchecked")
-    @DeleteMapping("/{idUser}/tasks/{idTask}")
+    @DeleteMapping("/user/{idUser}/task/{idTask}")
     public Map<String, Object> deleteTaskFromUser(@PathVariable("idUser") Long idUser, @PathVariable("idTask") Long idTask) {
         User user = userService.findUserById(idUser);
         if (user == null)
@@ -176,12 +171,11 @@ public class UserController {
         Task task = taskService.findTaskById(idTask);
         if (task == null)
             throw new NotFoundException("The task with idTask " + idTask + " does not exist.");
-        userService.removeTaskFromUser(user, task);
+        if (userService.getTasksFromUser(user).contains(task)) userService.removeTaskFromUser(user, task);
         return new ShowUser(user, userService.getShowTaskFromUser(user)).getFields(ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
 
-    @SuppressWarnings("unchecked")
-    @DeleteMapping("/{idUser}/tasks")
+    @DeleteMapping("/user/{idUser}/tasks")
     public Map<String, Object> deleteAllTasksFromUser(@PathVariable("idUser") Long idUser) {
         User user = userService.findUserById(idUser);
         if (user == null)
@@ -191,7 +185,7 @@ public class UserController {
     }
 
     @SuppressWarnings("unchecked")
-    @PutMapping("/{idUser}/token")
+    @PutMapping("/user/{idUser}/token")
     public Map<String, Object> updateToken(@PathVariable("idUser") @Min(value = 0, message = "The idUser must be positive.") Long idUser, ServletWebRequest request) {
 
         User user = userService.findUserById(idUser);
@@ -206,18 +200,14 @@ public class UserController {
         return new ShowUser(user, userService.getShowTaskFromUser(user)).getFields(ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
 
-    @SuppressWarnings("unchecked")
-    @GetMapping("/tasks/{idTask}")
+    @GetMapping("/users/task/{idTask}")
     public List<Map<String, Object>> getUserWithTask(@PathVariable("idTask") @Min(value = 0, message = "The idTask must be positive.") Long idTask) {
         Task task = taskService.findTaskById(idTask);
         if (task == null)
             throw new NotFoundException("The task with idTask " + idTask + " does not exist.");
-        Preconditions.checkNotNull(task, "The task with idTask " + idTask + " does not exist.");
         List<User> users = userService.findUsersWithTask(task);
         if (users == null || users.isEmpty())
             throw  new BadRequestException("The task with idTask " + idTask + " does not belong to any user.");
         return users.stream().map(user -> new ShowUser(user, userService.getShowTaskFromUser(user)).getFields(ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES)).toList();
     }
-
-
 }
